requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
boto==2.49.0
boto3==1.9.130
botocore==1.12.130
bz2file==0.98
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
docutils==0.14
gensim==3.7.2
idna==2.8
jmespath==0.9.4
mecab-python3==0.996.1
numpy==1.16.2
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
s3transfer==0.2.0
scipy==1.2.1
six==1.12.0
smart-open==1.8.1
soupsieve==1.9.1
tqdm==4.31.1
urllib3==1.24.1
